+++ src/containers/group/detail/GroupDetail.js
